services/muse-agentic-rag/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph>=0.0.30
langchain>=0.1.0
langchain-core>=0.1.0
tavily-python>=0.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
google-generativeai>=0.4.0
sse-starlette>=1.6.0
pydantic>=2.0.0
httpx>=0.25.0
numpy>=1.24.0
groq>=0.4.0
